Add mad_libs.rb
Create file to complete homework assignment. It was a pain to
format the text paragraph as EXACTLY specified in the
step_definitions directory, but I finally got it done.

In order to avoid warning from Rubocop, I created a list of
variables to pass in strings to the get_word method.

[#105164714]