Modify my_object.rb to fix rubocop erros and clean code.
SassyOven is now devoid of offenses detected by rubocop and has
code that is easy to read and interpret. I made a new file called
my_object_auxiliary_methods.rb to contain methods that are
essential to the function of SassyOven, but that the user doesn't
need to call. So I hid them from the user all together. I may have
silenced Rubocop in some sections, but only because it thought
that my methods/lines were too long and I knew better.

[#105361970]